pyproject.toml
[project]
name = "health-up"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "alembic>=1.15.1",
  "fastapi[standard]>=0.115.8",
  "kafka-python>=2.0.5",
  "passlib[bcrypt]>=1.7.4",
  "psycopg2>=2.9.10",
  "pyjwt>=2.10.1",
  "python-decouple>=3.8",
  "sqlalchemy-utils>=0.41.2",
  "sqlmodel>=0.0.23",
  "uuid7>=0.1.0",
]
